+++ CHANGELOG.md
@@ -14,7 +14,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 - Adds two patterns for rendering the updated `Modal` component. This is still an ongoing work-in-progress as the `Modal` gets finalized but it is now using Chakra under-the-hood.
 - Adds `@chakra-ui/storybook-addon` so stories automatically pick up the Chakra-based NYPL theme.
-- Adds the `SkipNav` component and adds this component inside the `TemplateAppContainer` component. The `SkipNav` component renders accessible links to the main content of a page and accessibility information on NYPL.org.
+- Adds the `SkipNavigation` component and adds this component inside the `TemplateAppContainer` component. The `SkipNavigation` component renders accessible links to the main content of a page and accessibility information on NYPL.org.
 
 ### Updates